ex05a/requirements.txt
python-dotenv
cassandra-driver
prettytable
PyYAML
